Refactor Datastore & Relationships
